fix(audio): widen Swift zone rootMidi to Double so tuneCents folds exactly
Swift's SampleZoneData.rootMidi was Int, so buildZone rounded the folded
root (rootMidi + tuneCents/100) to the nearest semitone, making tuneCents
a no-op on Swift (up to 50 cents of pitch error) — the previous "sanctioned
asymmetry" was wrong; fine-tuning is exactly what 3b's real piano work
needs. Widen rootMidi to Double (Pitch gains a Double overload, PackLoader
folds tuneCents exactly, nearestZone's tie-break casts midi to Double and
keeps the TS twin's lower-zone tie preference). Confirmed numerically inert
for all-integer roots: golden sample render and SampleZoneGeneratorTests
pass unchanged.

Also: add BasePathPackSource test coverage (fetchManifest request URL +
parsed result, invalidManifest throw path, fetchZone passthrough), drop
TS PackLoader's dead `manifest` field, and make Swift's buildZone only
carry loop points when both loopStart and loopEnd are set, matching TS.

+++ swift/Sources/AlloyAudio/DSP/SampleZoneGenerator.swift
@@ -3,13 +3,13 @@ import Foundation
 /// Sample playback with zones, velocity layers, loops, and Catmull-Rom
 /// interpolation. Twin of web src/dsp/sample-zone-generator.ts (canonical).
 public struct SampleZoneData {
-    public let rootMidi: Int
+    public let rootMidi: Double
     public let sampleRate: Double
     public let data: [Float]
     public let loopStart: Int?
     public let loopEnd: Int?
 
-    public init(rootMidi: Int, sampleRate: Double, data: [Float], loopStart: Int? = nil, loopEnd: Int? = nil) {
+    public init(rootMidi: Double, sampleRate: Double, data: [Float], loopStart: Int? = nil, loopEnd: Int? = nil) {
         self.rootMidi = rootMidi
         self.sampleRate = sampleRate
         self.data = data
@@ -130,8 +130,8 @@ public final class SampleZoneGenerator: ToneGenerator {
     private static func nearestZone(_ zones: [SampleZoneData], midi: Int) -> SampleZoneData {
         var best = zones[0]
         for zone in zones {
-            let d = abs(zone.rootMidi - midi)
-            let bestD = abs(best.rootMidi - midi)
+            let d = abs(zone.rootMidi - Double(midi))
+            let bestD = abs(best.rootMidi - Double(midi))
             if d < bestD || (d == bestD && zone.rootMidi < best.rootMidi) {
                 best = zone
             }
